Add Spaces Component to the Repo (#999)
Spaces is a grid layout component that allows users to layout a custom arrangement of components. It's a prototype model for a layout that is not an editor. It also plays with templating and how document type construction could work.

It's also the first component to use the new ContainerServices in Aqueduct. It was amazingly simple to move my existing Manager component to a service. Took about 2 minutes.

It's still pretty rough and needs a lot more work but is in a good enough state to get it in so we can iterate. 🙂

You can play with it here:
https://www.wu2.prague.office-int.com/waterpark/chemical-rest?chaincode=@fluid-example/spaces@0.13.0-skjokiel-test-0