Add MatchMaker Container Service into Vltava (#1093)
This change adds a new Container Service to the Aqueduct called the MatchMaker. This is more or less a port of Registro from the bohemia repo.

While doing Vltava I quickly ran into the scenario of wanting to have the Producer/Consumer model. Instead of writing a simple, and bad, version I wanted to leverage the existing work in the space. This lead me to writing an implementation off the existing interfaces. This quickly made me realize I was just re-writing what was probably in Registro.

This currently lives in the Aqueduct because I was planning on making it a first party citizen on the PrimedComponent. Since it uses some pieces from the Aqueduct if I made it its own package there would be some circular dependencies. I decided to not do that work just yet but instead write a bunch of tests to validate it.

MatchMaker is designed to be a Container Service that allows loaded/initialized components as well as other ContainerServices to discover capabilities. It's very very generalized based off the interfaces we have which makes it very flexible but also not really intuitive to use.

I'll be integrating this into Vltava and maybe a few other components after this change gets in.

The one thing I learned from implementing these interfaces is that discover and discoverable is very very confusing syntax. However, I wasn't making it a goal to overhaul the interfaces, since there was a lot of time, effort, and discussion put into them already. We can discuss how to change them later if we want to. They seem pretty functional and more or less in parity with what bohemia has and is working with.