Fix fluid debugger assert (#1094)
* Fix assert when closing out of fluid debugger

Fixes #787

When loading a document with fluid debugger, it presents a popup asking to select a version to load, where if the popup is closed the document should load as usual.  If the popup is closed though, the debugger asserts.  This is because the process of creating the document service calls initialization on the storage service, but it's a shared storage service and its initialzation produces side effects.  Multiple document services are created, so we should only run the initialization once.

* iteration updates

Move responsibility for tracking and preventing multiple initializations into fluidDebuggerController rather than in replayDocumentService.

* Update fluidDebuggerController.ts