Vltava - Dynamic Component Discovery via Registry (#1126)
This PR explores the Container Registry as a discovery mechanism for the Tabs Component.

Before this change the Tabs Component hard coded all it's components and it was required that the Tabs Component brought these components along or they were defined in the container.

Now, when constructing the Container we pass in a Registry that we treat as an IComponent. Components can request the registry and query for the available components and capabilities of that components. This is currently schematized and pretty specific to Vltava but This can definitely be extensible. Now when new components are added to the Container Registry these components will be available for other components to consume.

This isn't an end goal but a good stopping place in the progression.

Example:

https://www.wu2.prague.office-int.com/waterpark/damaged-sister1?chaincode=@fluid-example/vltava@0.13.0-skjokiel-1

Also in this pr
Added FacePile to the top nav bar with piped users.