Update Riddler to store historian endpoint
Currently the service relies on the historian to abstract access to all underlying storage providers. But this proves to be limiting since it assumes a single historian service.

By maintaining the historian endpoint within the tenant configuration data the service can be configured to use a different historian than the default one that comes with the cluster.

Longer term we will update the service to use a driver based model for access to Historian rather than assuming it implements the git REST api.