Wrap component div in shadow
How many divs is too many divs?  The nav-bar styling for waterpark messes up the styling of the inserted components, so wrap all the components in a shadow dom.  The div holding the components comes from an hjs file though and it's not really possible to represent a shadow-root in plain html, so just dump everything in an extra div