Step1 of making hashing async (#985)
* Step1 of making hashing async

Partially fixes #746, followup items are linked

Blob hashing is happening synchronously, which can disrupt responsiveness to the user.  Its results aren't needed immediately, so it can be done asynchronously.

Step 1: Add an async version of the API in question that can use browser native APIs and update direct consumers to use the async version.  Updating package consumers under will happen in a later step (package consumers will need to wait until the changes are available in a newer version of the package).  Continue to use the git hashing algorithm with SHA1 because it needs to be compatible with git in some locations.  This involves an extra buffer copy when using the native browser hashing APIs because they don't support incremental hashing, but the speedups from using native code outweigh this.

Step 2 (later): Update package-consuming code to use the new async API once it has propagated to a new package version.