Fluid-Build: Make to manage all the packages
With monorepo split fluid-build was managing the monorepo separately.
But with th emove of common build and library packages, we want
fluid-build to manage the build more holistically.

This change make Fluid-build aware of all the package at all times.
And use switches to filter which set of package to build.

Make sure package install to work on common packages as well.

Improve --symlink and build/clean to beware of cross dependencies and
link and build them as needed as well.

Add sementic version checking to see of these dependencies are actually
related.