Remove usage of addView on the Math component in flow-scroll (#1079)
flow-scroll/webflow now create MathViews without routing through the MathInstance.  It does this by storing a view indicator string on the marker representing the MathView, and matching this to a view factory passed in a view registry from the host.  This contract of view indicators mapping to view factories becomes part of the public interface of webflow to be used by its hosts.