Changes to get close/stop signal down from container to component
Stop propagating an event from the container level down to the component and use direct calls to close/stop/dispose methods instead (and generally prefer concrete lifecycle methods over events).  Do this by implementing IDisposable in the places that need it.  This leaves some extra close/stop methods that should be cleaned up later because they kind of duplicate the same functionality as dispose().

Also convert a few stop functions to sync that don't need to be async, and split out snapshotting from close in cases that do both.