Add HTMLViewAdapter and use it in places where we abstract into IComponentHTMLViews (#1308)
In various places throughout our repo, we have replicated logic to inspect an IComponent and try to ultimately pull an IComponentHTMLView from it for rendering. This change adds an HTMLViewAdapter which consolidates this logic, and also adds logic to handle abstracting React components as well.

This adapter is then used in scenarios that were abstracting between IComponentHTMLView and IComponentHTMLVisual.

This change also renames EmbeddedComponent to ReactViewAdapter to be more revealing about its purpose.