Get TableView working in external component loader
Initially broken because external component loader is unable to find table document.  This is because the way the component registries are currently set up, both table view and table document are under the table view chaincode.  The table view component is the one that creates the table document, so table document should be in table view's component registry.  Further, the createAndAttachComponent API doesn't set up subregistries like the context.createComponent API does, and causes external component loader to look in its own registry when trying to find table document.