Table view xcl (#1435)
* Get TableView working in external component loader

Initially broken because external component loader is unable to find table document.  This is because the way the component registries are currently set up, both table view and table document are under the table view chaincode.  The table view component is the one that creates the table document, so table document should be in table view's component registry.  Further, the createAndAttachComponent API doesn't set up subregistries like the context.createComponent API does, and causes external component loader to look in its own registry when trying to find table document.

* Update externalComponentLoader.ts

add table-view as a default component in external component loader

* Update shared component to use new API

Move the fix to use the subregistry-aware createcomponent API from tableview into createAndAttachComponent so that existing users will get the fix as well (wasn't sure about the intent in the earlier commit so I scoped the change more narrowly)