Pass container close events down to SharedComponent (#1313)
* Pass container close events down to SharedComponent

Provide a way for components to run code on container close by registering for the event upward from the component to the container.  Also change the event name in the container from "closed" to "close" to normalize on typical naming convention for event names (present tense rather than past tense)*

*Some of the other events are still in the past tense here; is that fine or should all/none of them be changed?

* undo changes to events and use lifecycle methods

Use lifecycle methods over lifecycle events because we don't need the indirection of events here.  For some points, this means implementing IDisposable.  For others (i.e. the Runtimes which implement async stop method), hook into the stop methods instead because they are functionally the same.  A future change would add IAsyncDisposable and use that over stop.

* Changes to get close/stop signal down from container to component

Stop propagating an event from the container level down to the component and use direct calls to close/stop/dispose methods instead (and generally prefer concrete lifecycle methods over events).  Do this by implementing IDisposable in the places that need it.  This leaves some extra close/stop methods that should be cleaned up later because they kind of duplicate the same functionality as dispose().

Also convert a few stop functions to sync that don't need to be async, and split out snapshotting from close in cases that do both.

* Update sharedComponent.ts

* addressing comments

* remove idisposable from icontainercontext into containercontext

forgot this one earlier

* Update container.ts

* iteration changes

* Update componentContext.ts

* Update inframehost.ts

new class implementing iruntime

* fix context snapshotting

There was a conflictless merge around snapshotting that my refactor ended up missing.  Add a helper function in containterContext that handles snapshotting runtime state (instead of doing it through a stop method) and use that instead of creating the runtimestate object in the container because the containerRuntime isn't exposed through the containerContext (which is needed for the runtime snapshot)