Integrate package.json sorting into repo-policy-check (#1598)
Currently the CI runs the check of whether package.json is sorted at lint time.
However, CI always use lerna version to update the package version, which implicitly sort the package.json, so the check in the CI doesn't really detect if the file checked in is not sorted.

Move the sort check into repo-policy-check, which already scan all the package.json that is in the repo (and the untracked package.json as well) for author/license field. The CI run this separately with a pristine repo on every PR. And this change also expand the check for all package.json in the repo.