Docs: major infrastructure updates (#1630)
## Summary of changes

### Support for "variable groups" in the docs system

This adds support for "variable groups" in the Fluid docs system. There are two variable groups currently: _internal_ and _external_. We can now **add Markdown content that is shown/hidden per group**, as shown below.

### Add support for conditional sections in markdown content

In a Markdown file, wrap the conditional content block in the following:

```
<vue-markdown v-if="$themeConfig.fluidVarGroup === 'internal'">
### Some markdown content

This will only render if the variable group is internal.
<vue-markdown />
```

There are examples of this in the markdown files within this PR.

Using the `$themeConfig` variable is the only was I could find to reliably get a value from within a markdown file for the purposes of conditional logic. I don't like it, but it works!

### Vary the nav and content based on variable groups

The variable group at build time is defined in docs/.vuepress/config.js as:

```js
const fluidVarGroup = process.env[`FLUID_VAR_GROUP`] || "internal";
```

THe config dynamically builds the nav and content with custom code that lives in the config file. This code could be better organized and factored, but the patterns are pretty easy to use as we add more content even without that cleanup.

### NPM scripts for building external docs

These scripts can be used to manually run the docs build for the _external_ variable group. It does this by setting the `FLUID_VAR_GROUP` env variable before running the build scripts. In addition, it **deletes content in the docs folder** so it's omitted from the build output. **These changes should not be committed.**

### Misc other stuff

* Disabled the PWA since it has problems with the auth system
* Move include files into .vuepress so they aren't output by the build as individual pages
* Added a preliminary "build all the docs in the repo" script to help with debugging
* Ported some content from the dev preview docs
* Updated some deps