Update factory instantiation to use type param
Update all Shared/PrimedComponentFactory users to provide a name/type param.  For most users, the name was already defined whereas for some it wasn't already exposed.  There are some cases where the name being defined/used with the component factory doesn't match that used when creating the SimpleModuleInstantiaionFactory.  This doesn't appear to be an issue right now, but should be cleaned up later to be more consistent.

Also undo some earlier changes to add a template type to the factories.  While this allows stronger typing on the factory's CreateComponent call, it creates difficulties when defining component registies for the module instantiation factory.  This is because the typing for the map implicitly becomes <string, PrimedComponentFactory<Component1> | PrimedComponentFactory<Component2> | ...> rather than <string, PrimedComponentFactory<IComponent & IComponentLoadable>>.  Explicitly typing to the latter works, but it's messy and confusing and IMO worse than retyping the return value from CreateComponent (adding a genertic to the CreateComponent method like the existing API is an option as well)