Bug fix for incorrect null check on gateway guest user access.
After migrating to strict null checks an incorrect conditional was introduced that checked if the guest account was defined before creating it. As opposed to checking if the guest account was undefined.

While making the fix also rev'd to latest stable runtime and server bits.

And added a docker-compose mode that runs r11s without Gateway. And a compose file for Gateway that can bind to a running r11s cluster. These make it a little easier to deal with and combine the two services.