Add initial state flow into factory component creation
Add an optional initialState arg when creating components through the factory.  This is similar to the existing props args in the old component creation flow.  This does not grant type safety out of the box because the initialState is an any,  but it allows the component creator to extend Primed/SharedComponentFactory to override the createComponent method to specify a more explicit initial state type, which then provides type safety to any consumers who have the specifc factory itself.

I updated the ClickerWithInitialValue example to use this path (which appears to be the only consumer of the old path anyway?)