Fix MergeTree Client startCollaboration to Support Detached Container (#1747)
There was a bug in the client startCollaboration where we would assign a client id if one didn't exist. This is not correct and is breaking detached container, as the merge tree isn't staying in local mode until the container is attached. I've combined start and update collaboration to make it easier to call. Now if client id is undefined we will not enter collaboration mode on the merge tree. For detached mode we will enter collaboration once the container is attached an connected.

* Fix MergeTree Client startCollaboration to Support Detached Container

* fix packages

* revert unecessary change

* fix comment typos

Co-authored-by: Tony Murphy <anthonm@microsoft.com>