Remove initial state from public interfaces
Remove initial state args from component/componentfactory interfaces because we do not want them to be supplied generically - users must have access to the specific factory type in order to do so.

This requires a bit of indirection to workaround never allowing the initialState to be used as an `any`.