Add dispose event to container runtime
Fixes #1712
Contianer services don't directly register with the container runtime and instead hook in through events.  Container presence manager wants to do something on container runtime dispose, but the runtime can't call dispose on the presence manager, so we add the "dispose" event.

Also add "on" declarations for all of ContainerRuntime's emitted events for the listener type chex