Allow component initial state in factory component creation path (#1759)
Add internal hooks to SharedComponentFactory so consumers can extend and call these hooks to allow providing initial state with strong typing.  There are no points inside SharedComponentFactory that expose initial state hooks to generic factories (so consumers must have the specific factory type to provide initial state), nor is the initial state ever moved around generically (so types are enforced).  ClickerWithInitialValue is updated first as an example.

* Add initial state flow into factory component creation

Add an optional initialState arg when creating components through the factory.  This is similar to the existing props args in the old component creation flow.  This does not grant type safety out of the box because the initialState is an any,  but it allows the component creator to extend Primed/SharedComponentFactory to override the createComponent method to specify a more explicit initial state type, which then provides type safety to any consumers who have the specifc factory itself.

I updated the ClickerWithInitialValue example to use this path (which appears to be the only consumer of the old path anyway?)

* Remove initial state from public interfaces

Remove initial state args from component/componentfactory interfaces because we do not want them to be supplied generically - users must have access to the specific factory type in order to do so.

This requires a bit of indirection to workaround never allowing the initialState to be used as an `any`.

* iteration feedback

* move factory

move factory to remove warnings