Fix Shared Text Test (#1795)
This test had a couple problems. It was including divs when getting the text, which included presence information, and it wasn't waiting for the document to converge. I ran the old version is a loop for 100 iterations which reproduced the failure. After the fix the test passed all 100 iterations.

Fixes #1752