Fix Local Server Summaries (#1801)
Local server was using and orderer manager that does support summaries correctly because it never started scribe. This change takes the order manager from tinylicous, which does support scribe. This enables summaries in local-server. tested with symlink:full. On this is published we can update tinylicious to use it from this package

fixes #1001
related #981