Always Specifiy ClientId on Load From Snapshot (#1806)
in the change for startOrUpdateCollaboration i missed a case where we load from snapshot/attach message, in which case we need to start collaboration on the merge tree before we have a client id to apply outstanding ops. in that case we'll got back to the old behavior and specific a default client id until our real one shows up.