initial commit
Try adding some generics to the component factories to reduce the amount of boilerplate code that components need to provide in order to hook into using initial state creation.  This may interfere with the other generics being added for typed events and component providers, will need to be resolved.  Also add a registryEntry property for the factories that types the factory to the lowest common denominator generic types, because ts doesn't infer this when mixing factories with generics.

Update a few examples