Fluid Synthesize integration in Aqueduct (#1812)
This change adds the new Synthesize package into the Aqueduct.

We can definitely iterate on this but I'm really liking this development model.

Note: It touches a lot of files because this integration requires updating the signature of the PrimedComponentFactory.

High Level What's in the change
Integration of Synthesize package with Aqueduct
Adding an example to Pond
What are the actual changes
Component Developers can now define optional providers that will be available on the this.providers object.

Container Developers using the Base Container can now supply providers. These providers will be injected to Component developers that ask for it

There are changes to Synthesize to make the developer experience better.

Pond Changes
There is a new UserInfo Provider that uses the quorum to provide user information about the container. The Pond container injects this service

There is a new component ExampleUsingProviders in Pond that leverages the UserInfo Provider.