Hotfix testhost to store components on its own root (#1959)
Bohemia doesnt have a DDS to store the handles on. As such, it needs to have the test root handle storing and retrieving its components.

ScriptorPad is using TestHost and they don't have a DDS of their own. This is a workaround for ScriptorPad still using ID while we have removed IDs in 0.16. Ultimately, we should migrate ScriptorPad off of using TestHost as they are treating it as a container but that is a bigger change and should be isolated from the version migration.

fixes #1962 and #1628