Send and verify a nonce on connect (#1919)
* Send and verify a nonce on connect

* Explicitly check against undefined

* Linting

* Update lerna-package-lock.json

* fix PR build issues

If more than one client tries sending connect_document on the multiplexed websocket before a connection handshake finishes, they can get confused by another client's message. This adds a UUID nonce to the request, and the given document will ignore responses that don't have a matching nonce (support for this on the Push side is ready to go when this is approved)

Backwards compatibility is maintained via not performing the check if the client or the server don't have a nonce.