refactoring
Refactor index again so that that the entry is always the same file, but the browser key replaces a file that's included in index that includes all platform specific files.  This means we never need to modify the main or module keys, and browser only ever has at most 2 entries.