Add event for when components are instantiate. (#1778)
Bohemia wants a way to log telemetry when new components are instantiated. I'm adding this event that will emit when ever a new component has been instantiated. It currently provides three bits of information:

Component Package Name
Registry Path
If it's a create new or not
We are changing the way components are created and retrieved and it will break the way they log. This is designed to be a light weight way to plug into.

I'm not a huge fan of this change but the way we load components is so async and pretty spread out that it's hard to tie into.

The other thing I could do is use the fact that IHostRuntime is an EventEmitter to emit an event directly on the ComponentRuntime but I hate this pattern because it's way to invasive.

If you have thoughts on how to make this better let me know.