Add Support for Body Blob Array in MergeTree Snapshot and clean V1 Contract (#1904)
* add chunk v1

* support chunk arrays

* fixes

* readd start index

* rename snapshot files to legacy

* revert some uncessary changes

* add new format tests

* get rid of intermediate version

* more simplification

* add copyright header

* integrate v1 snapshot tests

* lint fixes

* delete old files

Co-authored-by: Tony Murphy <anthonm@microsoft.com>

This is an extension of the work @DLehenbauer did to support snapshotting blue segments.

This change adds backwards and forwards support for Body Blob Arrays in the merge tree, in addition to backwards and forward support for a cleaned up contract in V1. We need the content blobs to be an array to support more advanced virtualization, and stable partitioning in the future. Adding the arrays now will ease in keeping those change backwards compatible, as older clients can ignore anything else added, and just serially load the blobs like they do now.

Move all chunk related contracts and helper functions to snapshotChucks file
Changed loader to always work with the latest contract regardless of contract used in snapshot
Snapshot Legacy uses the old format, but also add necessary headerMetadata so once all clients have this code we can move to the new format.
Snapshot V1 cleaned up the contract
sequenceTestSnapshot*.json existing files now have header metadata, but are still backwards compatible with clients that don't know about arrays. The new files are in the new format.