We may send same ops multiple times due to view-only connection (#2037)
Issue #2007: DeltaManager should not reconnect as view-only if there were any pending ops at the moment of last disconnect. See issue for more details.

All of this logic should likely move down to container runtime, and deltaManager ask it for reconnection mode instead. But that's left for another day.

Somewhat unrelated changes in containerRuntime.ts, based on a fact that code in processInboundMessage() was wrong regarding using connection clientId, which lead me to look at tracking of dirty state.