Improve Manual Attach in Webpack Component Loader (#2078)
Move the manual attach flow to be programmatic, to remove the flash of the elements when not in use. Also simplify the flow by using the hash(#) in the url, now http://localhost:8080/#manualattach will show the button for triggering manual attach. the hash will be removed after the attach allowing refresh as normal