Ensure ConsensusOrderedCollection promises reject when ComponentRuntime disposes (#2066)
Fixes #1420, and provides an API to fix similar issues in other DDS's.

The idea is that any Promise created in a DDS that doesn't resolve until after an op has been ack'd will hang if the ComponentRuntime is disposed before the ack comes back. So I added a protected function newAckBasedPromise to SharedObject that will ensure the promise is rejected when the ComponentRuntime disposes.