stabalize jest test exec times
Jest tests kick off webpack-dev-server asynchronously at start, which takes a while to be ready, and is even slower when run under lerna. The wait for page load at the beginning of each jest test gets attributed to the first test run, which then leads to intermittent (or consistent on slower machines) timeouts.  It's not clear that webpack-dev-server can reuse previously built products because of how it runs in-memory, but we can move the initial wait for page load out from the first test and give it a larger timeout to stabilize the actual test runs.  Initial page load aside, the actual tests show very little variance in their run times, and we can also revert their timeouts to the default 5s.  Certain projects like sharedtext and spaces that had very large timeouts set previously needed them because their ordinal positioning usually put them at the end of the first batch of runs by lerna, causing them to take the longest to finish the webpack step.