Update r11s dockerfile to auth against VSTS and take the auth token as an environment param.
On the build server, due to a lack of support for _authToken in VSTS's private NPM feeds, we will use the .npmrc.vsts for auth and then overwrite .npmrc.

Also renames r11s-Dockerfile to Dockerfile and keeps the Docker context to be the same directory as the Dockerfile.

As part of this copies over the license and build tools. A subsequent checkin will make the build tools a npm project.