Change verifynotclosed flow
Previous change incorrectly changed the control flow of verifyNotClosed depending on dispose with error condition.  This changes it back so control flow does not change between the two conditions, but does so that we no longer throw an error and terminate whatever function in ContainerRuntime that called verifyNotClosed.

The alternative was to throw an error in all cases and selectively ignore it further up.  I think this approach is problematic because it would be unique in the codebase and consumers would need to account for this internal runtime behavior (there is no central point for this that is easily maintained).