Update r11s helm chart to be defined outside of a build script. (#2159)
We've learned a lot about Helm since we first started using it. Rather than checking in a build tool to update config files we instead use value overrides to change things like the container image. This avoids needing to rebuild it on every change. Instead we only need to rebuild when the chart definition actually changes.

Also adds updated build instructions to the readme.