Stop throwing in ContainerRuntime verifyNotClosed (#2141)
Previous change incorrectly changed the control flow of verifyNotClosed depending on dispose with error condition. This changes it back so we do not consider any error condition, and instead removes the check from methods that may legitimately be called after dispose due to asynchrony.  This is intended to mirror a readonly disconnected state.

In addition, do the same for ComponentRuntime and ComponentContext - remove the verifyNotClosed check for getting the quorum and audience, but keep it for methods that interface with the loader. For ComponentRuntime, also remove the stop() method and migrate it to use the disposed flag instead.