Fix Local Sever Performance Regression (#2151)
Local server was modified to use a shared memoryOrdererManager with tinylicious, but this brought with it a hack that delays orderer creation by 1 second. This change exposes the necessary promised off localOrderer so we can wait until the lambdas are started. I've also moved the logic for connecting to a websocket from the driver to local server. Once this change propagates we can change the driver to reused this implementation as well. The reason this was moved, was to enable testing.

fixes #2147