Add fallback for hashFile for 0.15 (#1950)
* Add fallback for hashFile for 0.15

Fixes #1922
SHA1 hash in subtle crypto fails on edge classic. Add a fallback for sublte crypto if it fails in 0.14, because release/0.15 consumes version 0.14 of common-utils
The affected API is hashFile, which is only used in packages projects. Server only uses gitHashFile. also fallback was incorrectly calling gitHashFile.  it produced consistent results on edge classic where the call always fails, but will not in any case it does not