Use PromiseCache in OdspCache for consistent and safe caching logic (#2240)
I wrote PromiseCache after seeing these and a few other places that are grappling with the subtle logic of trying to cache the result of async work, sometimes with expiration logic.

Each of the PromiseCaches here fixes a bug or clarifies a potential pitfall.  This also clarifies the scenario around ICache by renaming to IPersistedCache.