Add test-tools package for shared test support code (#2328)
Pt1 of fixing #1341

Add test-tools package to hold shared test support code, such as port management for jest tests (and maybe other things in the future?).

For jest ports, the issue is that they're currently hardcoded in the test configs, and all tests are run in parallel. Meaning that anyone adding a new test needs to find the next available port and hardcode that as well. This change adds a script that sets up a mapping of unique ports so that packages running tests will not see collisions when run under lerna (and new test writers do not need to futz around with port values).

Updating consumers coming in a followup after this new package becomes available for import (couldn't test some things such as the binary availability due to this)